requirements-dev.txt
# Development dependencies for OwnTracks Django Backend

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0

# Code quality
flake8>=6.1.0
black>=23.7.0
mypy>=1.5.0

# Django debug toolbar (optional)
django-debug-toolbar>=4.2.0
